Undo the refactor of `project_on_boundary`
