+++ src/tqec/plaquette/plaquette.py
@@ -2,13 +2,13 @@ from __future__ import annotations
 
 from collections import defaultdict
 from dataclasses import dataclass, field
-from typing import Literal, Mapping, Protocol
+from typing import Mapping 
 
 from typing_extensions import override
 
 from tqec.circuit.schedule import ScheduledCircuit
 from tqec.exceptions import TQECException
-from tqec.plaquette.enums import MeasurementBasis, ResetBasis
+from tqec.plaquette.enums import PlaquetteOrientation
 from tqec.plaquette.frozendefaultdict import FrozenDefaultDict
 from tqec.plaquette.qubit import PlaquetteQubits
 from tqec.position import Position2D
@@ -67,6 +67,39 @@ class Plaquette:
 
     def __hash__(self) -> int:
         return hash((self.qubits, str(self.circuit.get_circuit())))
+
+    def project_on_boundary(
+        self, projected_orientation: PlaquetteOrientation
+    ) -> Plaquette:
+        """Project the plaquette on boundary and return a new plaquette with the
+        remaining qubits and circuit.
+
+        This method is useful for deriving a boundary plaquette from a integral
+        plaquette.
+
+        Args:
+            projected_orientation: the orientation of the plaquette after the
+                projection.
+
+        Returns:
+            A new plaquette with projected qubits and circuit. The qubits are
+            updated to only keep the qubits on the side complementary to the
+            provided orientation. The circuit is also updated to only use the
+            kept qubits and empty moments with the corresponding schedules are
+            removed.
+        """
+        kept_data_qubits = self.qubits.get_qubits_on_side(
+            projected_orientation.to_plaquette_side()
+        )
+        new_plaquette_qubits = PlaquetteQubits(
+            kept_data_qubits, self.qubits.syndrome_qubits
+        )
+        new_scheduled_circuit = self.circuit.filter_by_qubits(
+            new_plaquette_qubits.all_qubits
+        )
+        return Plaquette(
+            new_plaquette_qubits, new_scheduled_circuit, self.mergeable_instructions
+        )
 
 
 @dataclass(frozen=True)
@@ -141,15 +174,3 @@ class RepeatedPlaquettes(Plaquettes):
 
     def __hash__(self) -> int:
         return hash((self.repetitions, super().__hash__()))
-
-
-class PlaquetteBuilder(Protocol):
-    """Protocol for functions building a `Plaquette`."""
-
-    def __call__(
-        self,
-        basis: Literal["X", "Z"],
-        data_qubits_initialization: ResetBasis | None = None,
-        data_qubits_measurement: MeasurementBasis | None = None,
-        x_boundary_orientation: Literal["HORIZONTAL", "VERTICAL"] = "VERTICAL",
-    ) -> Plaquette: ...
